// packages/api/openapi/clients/typescript/models/EventsListEventsLive200ResponseOneOf.ts
/**
 * Director API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { EventsListEventsLive200ResponseOneOfData } from '../models/EventsListEventsLive200ResponseOneOfData';
import { HttpFile } from '../http/http';

export class EventsListEventsLive200ResponseOneOf {
    'event': any | null;
    'data': EventsListEventsLive200ResponseOneOfData;
    'id'?: string | null;
    'retry'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "event",
            "baseName": "event",
            "type": "any",
            "format": ""
        },
        {
            "name": "data",
            "baseName": "data",
            "type": "EventsListEventsLive200ResponseOneOfData",
            "format": ""
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "string",
            "format": ""
        },
        {
            "name": "retry",
            "baseName": "retry",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return EventsListEventsLive200ResponseOneOf.attributeTypeMap;
    }

    public constructor() {
    }
}