// packages/api/openapi/clients/typescript/models/HealthLiveOk200ResponseOneOf.ts
/**
 * Director API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class HealthLiveOk200ResponseOneOf {
    'event': any | null;
    'data': string | null;
    'id'?: string | null;
    'retry'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "event",
            "baseName": "event",
            "type": "any",
            "format": ""
        },
        {
            "name": "data",
            "baseName": "data",
            "type": "string",
            "format": ""
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "string",
            "format": ""
        },
        {
            "name": "retry",
            "baseName": "retry",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return HealthLiveOk200ResponseOneOf.attributeTypeMap;
    }

    public constructor() {
    }
}