// packages/api/openapi/clients/typescript/models/OdienceValidatePhoneNumberRequest.ts
/**
 * Director API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class OdienceValidatePhoneNumberRequest {
    'msisdn': string;
    'countryCode'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "msisdn",
            "baseName": "msisdn",
            "type": "string",
            "format": ""
        },
        {
            "name": "countryCode",
            "baseName": "country_code",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return OdienceValidatePhoneNumberRequest.attributeTypeMap;
    }

    public constructor() {
    }
}